Create BasePreferenceFragment.kt with summary methods
